+++ views/ap/Friends.js
@@ -10,7 +10,7 @@ import { accept, reject } from './utils/postActivity'
 
 export default function Friends ({ iri }) {
   const [items, setItems] = useState([])
-  const { token } = useContext(ServerDataContext)
+  const { token, actor: me } = useContext(ServerDataContext)
   useEffect(() => {
     if (!token) {
       return
@@ -27,7 +27,7 @@ export default function Friends ({ iri }) {
   }, [iri, token])
   return (
     <div className='aesthetic-windows-95-container-indent'>
-      {items.map(item => <Friend key={item.id} {...item} />)}
+      {items.map(item => item.actor === me.id ? <PendingRequest key={item.id} {...item} /> : <Friend key={item.id} {...item} />)}
     </div>
   )
 }
@@ -65,4 +65,28 @@ function Friend ({ id, type, actor, summary, object = {}, target = {}, published
       </div>
     </div>
   )
+}
+
+function PendingRequest ({ id, type, actor, summary, object = {}, target = {}, published }) {
+  return null
+  // TODO: needs immers-client integration for removeFriend function
+  /*
+  const { token, actor: me } = useContext(ServerDataContext)
+  const [action, setAction] = useState('')
+  const { id: actorId, icon } = object
+  return (
+    <div className={c({ none: action === 'cancel' })}>
+      <div className='postHeader'>
+        <a className='handle profileLink' href={actorId}>
+          <ProfileIcon size='tiny' icon={icon} />
+          <ImmersHandle {...object} showName />
+        </a>
+        <span className={c('friendManager')}>
+          you sent a friend request:
+          <EmojiButton emoji='x' title='Cancel' onClick={() => { client.removeFriend(actorId); setAction('cancel') }} />
+        </span>
+      </div>
+    </div>
+  )
+  */
 }